# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Knowledge Agent Hub API
  version: 1.0.0
  description: Local-first evidence retrieval, knowledge management, and optional cited-answer API.
  contact:
    name: Knowledge Agent Hub maintainers
    url: https://github.com/KiriuKazurei/agent-knownledge-base
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:48761/api/v1
security:
  - bearerAuth: []
paths:
...
  /skills/{skillId}:
    parameters:
      - in: path
        name: skillId
        required: true
        schema: {type: string, format: uuid}
    get:
      operationId: getSkill
      responses:
        '200':
          description: Skill metadata
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Skill'}
    delete:
      operationId: deleteSkill
      responses:
        '204': {description: Deleted}
...
  /folders:
    get:
      operationId: listFolders
      parameters:
        - {in: query, name: libraryId, required: true, schema: {type: string}}
      responses:
        '200':
          description: Virtual folders
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/VirtualFolder'}}
    post:
      operationId: createFolder
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateFolderRequest'}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/VirtualFolder'}
  /folders/{folderId}:
    delete:
      operationId: deleteFolder
      parameters:
        - {in: path, name: folderId, required: true, schema: {type: string}}
      responses:
        '204': {description: Deleted}
  /documents/{documentId}/folders/{folderId}:
    put:
      operationId: assignDocumentFolder
      parameters:
        - {in: path, name: documentId, required: true, schema: {type: string}}
        - {in: path, name: folderId, required: true, schema: {type: string}}
      responses:
        '204': {description: Assigned}
    delete:
      operationId: removeDocumentFolder
      parameters:
        - {in: path, name: documentId, required: true, schema: {type: string}}
        - {in: path, name: folderId, required: true, schema: {type: string}}
      responses:
        '204': {description: Removed}
  /sources/watches:
    get:
      operationId: listSourceWatches
      parameters:
        - {in: query, name: libraryId, required: true, schema: {type: string}}
      responses:
        '200':
          description: Source watches
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/SourceWatch'}}
    post:
      operationId: createSourceWatch
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateSourceWatchRequest'}
      responses:
        '202': {description: Scan queued}
  /sources/watches/{watchId}:
    delete:
      operationId: deleteSourceWatch
      parameters:
        - {in: path, name: watchId, required: true, schema: {type: string}}
      responses:
        '204': {description: Deleted}
  /sources/watches/{watchId}/scan:
    post:
      operationId: scanSourceWatch
      parameters:
        - {in: path, name: watchId, required: true, schema: {type: string}}
      responses:
        '202': {description: Scan queued}
  /indexes/rebuild:
    post:
      operationId: rebuildIndex
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RebuildIndexRequest'}
      responses:
        '202': {description: Rebuild queued}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  schemas:
...
    Skill:
      type: object
      required: [id, name, description, rootPath, entryPoint, contentHash, status, fileCount, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        name: {type: string, maxLength: 64}
        description: {type: string, maxLength: 1024}
        compatibility: {type: string, maxLength: 500}
        license: {type: string}
        metadata: {type: object, additionalProperties: {type: string}}
        allowedTools: {type: string}
        rootPath: {type: string}
        entryPoint: {type: string, const: SKILL.md}
        contentHash: {type: string}
        status: {enum: [ready, failed, invalid]}
        error: {type: string}
        systemRole: {type: string}
        fileCount: {type: integer, minimum: 0}
        usesLibraryIds: {type: array, items: {type: string, format: uuid}}
        requiresLibraryIds: {type: array, items: {type: string, format: uuid}}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
...
    VirtualFolder:
      type: object
      required: [id, libraryId, name, createdAt, updatedAt]
      properties:
        id: {type: string}
        libraryId: {type: string}
        name: {type: string}
        parentId: {type: string}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    CreateFolderRequest:
      type: object
      required: [libraryId, name]
      properties:
        libraryId: {type: string}
        name: {type: string, minLength: 1, maxLength: 120}
        parentId: {type: string}
    SourceWatch:
      type: object
      required: [id, libraryId, rootPath, recursive, enabled]
      properties:
        id: {type: string}
        libraryId: {type: string}
        rootPath: {type: string}
        recursive: {type: boolean}
        enabled: {type: boolean}
        lastScanAt: {type: string, format: date-time}
        lastMessage: {type: string}
    CreateSourceWatchRequest:
      type: object
      required: [libraryId, rootPath]
      properties:
        libraryId: {type: string}
        rootPath: {type: string}
        recursive: {type: boolean, default: true}
    RebuildIndexRequest:
      type: object
      required: [libraryId]
      properties:
        libraryId: {type: string}
...